migrated 2 files from js to ts
