+++ src/platform/web/ui/session/RoomGridView.js
@@ -17,7 +17,7 @@ limitations under the License.
 import {RoomView} from "./room/RoomView.js";
 import {InviteView} from "./room/InviteView.js";
 import {TemplateView} from "../general/TemplateView";
-import {StaticView} from "../general/StaticView.js";
+import {StaticView} from "../general/StaticView.ts";
 
 export class RoomGridView extends TemplateView {
     render(t, vm) {
